// src/payload/generated-schema.ts
export const media = pgTable(
  'media',
  {
    id: uuid('id').defaultRandom().primaryKey(),
    alt: varchar('alt'),
    prefix: varchar('prefix').default('media'),
    updatedAt: timestamp('updated_at', {
      mode: 'string',
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
    createdAt: timestamp('created_at', {
      mode: 'string',
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
    url: varchar('url'),
    thumbnailURL: varchar('thumbnail_u_r_l'),
    filename: varchar('filename'),
    mimeType: varchar('mime_type'),
    filesize: numeric('filesize', { mode: 'number' }),
    width: numeric('width', { mode: 'number' }),
    height: numeric('height', { mode: 'number' }),
    focalX: numeric('focal_x', { mode: 'number' }),
    focalY: numeric('focal_y', { mode: 'number' }),
  },
  (columns) => [
    index('media_updated_at_idx').on(columns.updatedAt),
    index('media_created_at_idx').on(columns.createdAt),
    uniqueIndex('media_filename_idx').on(columns.filename),
  ],
);
// ...
export const brands = pgTable(
  'brands',
  {
    id: uuid('id').defaultRandom().primaryKey(),
    title: varchar('title').notNull(),
    slug: varchar('slug').notNull(),
    slugLock: boolean('slug_lock').notNull().default(true),
    srcImage: varchar('src_image'),
    image: uuid('image_id').references(() => media.id, {
      onDelete: 'set null',
    }),
    srcUrl: varchar('src_url'),
    active: boolean('active').notNull().default(true),
    updatedAt: timestamp('updated_at', {
      mode: 'string',
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
    createdAt: timestamp('created_at', {
      mode: 'string',
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
  },
  (columns) => [
    uniqueIndex('brands_title_idx').on(columns.title),
    index('brands_slug_idx').on(columns.slug),
    index('brands_image_idx').on(columns.image),
    index('brands_updated_at_idx').on(columns.updatedAt),
    index('brands_created_at_idx').on(columns.createdAt),
  ],
);

export const products = pgTable(
  'products',
  {
    id: uuid('id').defaultRandom().primaryKey(),
    active: boolean('active').default(true),
    brand: uuid('brand_id')
      .notNull()
      .references(() => brands.id, {
        onDelete: 'set null',
      }),
    title: varchar('title').notNull(),
    slug: varchar('slug').notNull(),
    slugLock: boolean('slug_lock').notNull().default(true),
    origin: varchar('origin'),
    srcImage: varchar('src_image'),
    productInformation: varchar('product_information'),
    nutritionalInformation: varchar('nutritional_information'),
    srcUrl: varchar('src_url'),
    updatedAt: timestamp('updated_at', {
      mode: 'string',
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
    createdAt: timestamp('created_at', {
      mode: 'string',
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
  },
  (columns) => [
    index('products_brand_idx').on(columns.brand),
    index('products_slug_idx').on(columns.slug),
    index('products_updated_at_idx').on(columns.updatedAt),
    index('products_created_at_idx').on(columns.createdAt),
  ],
);
// ...
export const product_lines = pgTable(
  'product_lines',
  {
    id: uuid('id').defaultRandom().primaryKey(),
    slug: varchar('slug').notNull(),
    slugLock: boolean('slug_lock').notNull().default(true),
    sku: varchar('sku').notNull(),
    active: boolean('active').notNull().default(true),
    title: varchar('title').notNull(),
    size: varchar('size'),
    flavorColor: varchar('flavor_color'),
    variant: varchar('variant'),
    product: uuid('product_id')
      .notNull()
      .references(() => products.id, {
        onDelete: 'set null',
      }),
    category: varchar('category'),
    srcPrice: numeric('src_price', { mode: 'number' }),
    srcDiscountedPrice: numeric('src_discounted_price', { mode: 'number' }),
    retailPrice: numeric('retail_price', { mode: 'number' }),
    wholesalePrice: numeric('wholesale_price', { mode: 'number' }),
    stock: numeric('stock', { mode: 'number' }).notNull().default(0),
    srcStock: numeric('src_stock', { mode: 'number' }),
    onPallet: numeric('on_pallet', { mode: 'number' }),
    inCase: numeric('in_case', { mode: 'number' }),
    expiryDate: timestamp('expiry_date', {
      mode: 'string',
      withTimezone: true,
      precision: 3,
    }),
    weight: numeric('weight', { mode: 'number' }),
    servingSizePerContainer: numeric('serving_size_per_container', { mode: 'number' }),
    servingSize: numeric('serving_size', { mode: 'number' }),
    srcProductCode: varchar('src_product_code'),
    barcode: varchar('barcode'),
    exportCommodityCode: varchar('export_commodity_code'),
    updatedAt: timestamp('updated_at', {
      mode: 'string',
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
    createdAt: timestamp('created_at', {
      mode: 'string',
      withTimezone: true,
      precision: 3,
    })
      .defaultNow()
      .notNull(),
  },
  (columns) => [
    index('product_lines_slug_idx').on(columns.slug),
    uniqueIndex('product_lines_sku_idx').on(columns.sku),
    index('product_lines_title_idx').on(columns.title),
    index('product_lines_product_idx').on(columns.product),
    index('product_lines_barcode_idx').on(columns.barcode),
    index('product_lines_updated_at_idx').on(columns.updatedAt),
    index('product_lines_created_at_idx').on(columns.createdAt),
  ],
);